WLTS-30 rename custom value package (part 2)
   [WLTS-30](https://tools.hmcts.net/jira/browse/WLTS-30)

Rename the package used by the custom value evaluators.
* Part 2: update references
