+++ src/functionalTest/java/uk/gov/hmcts/reform/translate/customvalue/EvaluatorUtils.java
@@ -1,9 +1,11 @@
-package uk.gov.hmcts.reform.translate;
+package uk.gov.hmcts.reform.translate.customvalue;
 
 import org.apache.commons.lang3.RandomStringUtils;
 
-public final class StringGenerator {
-    private StringGenerator() {
+public final class EvaluatorUtils {
+
+    // Hide Utility Class Constructor : Utility classes should not have a public or default constructor (squid:S1118)
+    private EvaluatorUtils() {
     }
 
     public static String generate() {
