+++ src/functionalTest/java/uk/gov/hmcts/reform/translate/customvalue/UniqueTranslationWithEnglishAndWelshEvaluator.java
@@ -1,9 +1,7 @@
-package uk.gov.hmcts.reform.translate.custom.value;
+package uk.gov.hmcts.reform.translate.customvalue;
 
 import uk.gov.hmcts.befta.player.BackEndFunctionalTestScenarioContext;
-import uk.gov.hmcts.reform.translate.StringGenerator;
 
-import java.sql.SQLOutput;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -16,12 +14,13 @@ public class UniqueTranslationWithEnglishAndWelshEvaluator implements CustomValu
 
     @Override
     public Object calculate(BackEndFunctionalTestScenarioContext scenarioContext, Object key) {
-        final String expectedValueStr = key.toString().replace("UniqueTranslationWithEnglishAndWelsh ", "");
+        final String expectedValueStr = key.toString()
+            .replace(CustomValueKey.UNIQUE_TRANSLATION_WITH_ENGLISH_AND_WELSH + " ", "");
         System.out.println("sdansdjaksjdkajs"+expectedValueStr);
         int uniquePhrasesAmount = Integer.parseInt(expectedValueStr);
         Map<String, String> map = new HashMap<>();
         for (int i = 0; i<uniquePhrasesAmount; i++) {
-            String englishPhrase = StringGenerator.generate();
+            String englishPhrase = EvaluatorUtils.generate();
             final String welshPhrase = englishPhrase + "-WELSH";
             map.put(englishPhrase, welshPhrase);
         }
