+++ src/functionalTest/java/uk/gov/hmcts/reform/translate/customvalue/UniqueTranslationWithOnlyEnglishEvaluator.java
@@ -1,7 +1,6 @@
-package uk.gov.hmcts.reform.translate.custom.value;
+package uk.gov.hmcts.reform.translate.customvalue;
 
 import uk.gov.hmcts.befta.player.BackEndFunctionalTestScenarioContext;
-import uk.gov.hmcts.reform.translate.StringGenerator;
 
 import java.util.Map;
 
@@ -14,6 +13,6 @@ public class UniqueTranslationWithOnlyEnglishEvaluator implements CustomValueEva
 
     @Override
     public Object calculate(BackEndFunctionalTestScenarioContext scenarioContext, Object key) {
-        return Map.of(StringGenerator.generate(), "");
+        return Map.of(EvaluatorUtils.generate(), "");
     }
 }
