+++ src/functionalTest/resources/features/F-002 - Put Dictionary/F-002_Verify.td.json
@@ -19,7 +19,7 @@
   "expectedResponse": {
     "_extends_": "Common_200_Response",
     "body": {
-      "translations": "${[scenarioContext][customValues][containsUniquePhrasesParentContext]}"
+      "translations": "${[scenarioContext][customValues][containsDictionaryFromContext]}"
     }
   }
 }
